added owner to all contacts

+++ model/index.js
@@ -1,8 +1,8 @@
 const Contact = require("./schemas/contact");
 
-const listContacts = async ({ limit = 5, page = 1, sortBy, sortByDesc }) => {
+const listContacts = async (userId, { limit = 5, page = 1, sortBy, sortByDesc }) => {
   const { docs: contacts, totalDocs: total } = await Contact.paginate(
-    {},
+    { owner: userId },
     {
       limit,
       page,
@@ -19,16 +19,16 @@ const listContacts = async ({ limit = 5, page = 1, sortBy, sortByDesc }) => {
   return { contacts, total, page: Number(page), limit: Number(limit) };
 };
 
-const getContactById = async (contactId) => {
-  const contactById = await Contact.findOne({ _id: contactId }).populate({
+const getContactById = async (userId, contactId) => {
+  const contactById = await Contact.findOne({ _id: contactId, owner: userId }).populate({
     path: "owner",
     select: "email subscription -_id"
   });
   return contactById;
 };
 
-const removeContact = async (contactId) => {
-  const contactToRemove = Contact.findByIdAndRemove({ _id: contactId });
+const removeContact = async (userId, contactId) => {
+  const contactToRemove = Contact.findByIdAndRemove({ _id: contactId, owner: userId});
   return contactToRemove;
 };
 
@@ -37,9 +37,9 @@ const addContact = async (body, userId) => {
   return newContact;
 };
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (userId, contactId, body) => {
   const updatedContact = await Contact.findByIdAndUpdate(
-    { _id: contactId },
+    { _id: contactId, owner: userId },
     body,
     { new: true }
   );
